change int on size_t in tests

+++ tasks/seq/vinicuk_t_complex_sparse_matrix_mult_csc/func_tests/main.cpp
@@ -48,8 +48,8 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_vectors) {
 
   mtrx_A.col_ptrs = {0, n};
   // mtrx_B.col_ptrs = {0, 1, 2, 3};
-  for (int i = 0; i <= n; i++) mtrx_B.col_ptrs.push_back(i);
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i <= n; i++) mtrx_B.col_ptrs.push_back(i);
+  for (size_t i = 0; i < n; i++) {
     mtrx_A.row_indexes.push_back((double)i);
     mtrx_A.values.push_back(std::complex<double>(1.0, 1.0));
     mtrx_B.row_indexes.push_back(0.0);
@@ -69,7 +69,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_vectors) {
   testTaskSequential.run();
   testTaskSequential.post_processing();
   std::complex<double> answer(0.0, 2.0);
-  for (int i = 0; i < n * n; i++) ASSERT_EQ(mtrx_res.values[i], answer);
+  for (size_t i = 0; i < n * n; i++) ASSERT_EQ(mtrx_res.values[i], answer);
 }
 
 TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_diganal_square_matrixs) {
@@ -104,7 +104,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_diganal_square_
   testTaskSequential.post_processing();
   std::complex<double> answer1(3.0, 4.0);
 
-  for (int i = 0; i < mtrx_res.values.size(); i++) ASSERT_EQ(mtrx_res.values[i], answer1);
+  for (size_t i = 0; i < mtrx_res.values.size(); i++) ASSERT_EQ(mtrx_res.values[i], answer1);
 }
 
 TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_triangle_square_matrix_on_vector) {
@@ -114,7 +114,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_triangle_square
   CSCComplexMatrix mtrx_res(1, n);
 
   int k = 1, p = 0;
-  for (int i = 0; i <= n; i++) {
+  for (size_t i = 0; i <= n; i++) {
     mtrx_A.col_ptrs.push_back(p);
     p += k;
     k++;
@@ -122,7 +122,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_triangle_square
 
   k = 1;
   p = 0;
-  for (int i = 0; i < mtrx_A.col_ptrs[n]; i++) {
+  for (size_t i = 0; i < mtrx_A.col_ptrs[n]; i++) {
     mtrx_A.values.push_back(std::complex<double>(2.0, 1.0));
     if (p >= k) {
       p = 0;
@@ -133,7 +133,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_triangle_square
   }
 
   mtrx_B.col_ptrs = {0, n};
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     mtrx_B.values.push_back(std::complex<double>(2.0, 1.0));
     mtrx_B.row_indexes.push_back(i);
   }
@@ -152,7 +152,7 @@ TEST(vinicuk_t_complex_sparse_matrix_mult_csc_seq, test_multiply_triangle_square
   testTaskSequential.post_processing();
 
   k = n;
-  for (int i = 0; i < mtrx_res.values.size(); i++) {
+  for (size_t i = 0; i < mtrx_res.values.size(); i++) {
     ASSERT_EQ(mtrx_res.values[i], std::complex<double>(3.0 * k, 4.0 * k));
     k--;
   }
